feat: add book validation to the /book/:id endpoint on deletion

+++ back-end/bloco-24-nodejs-orm-autenticacao/dia-1-orm-interface-da-aplicacao-com-o-banco-de-dados/index.js
@@ -66,6 +66,10 @@ app.post('/book/:id', async (req, res) => {
 app.delete('/book/:id', async (req, res) => {
   try {
     const { id } = req.params;
+
+    const book = await Book.findByPk(id);
+    if (!book) return res.status(404).json({ message: 'Livro não encontrado' });
+
     await Book.destroy({ where: { id } });
 
     res.status(200).json(book);
